requirements.txt
flask
Flask
Flask_Mail
mailjet_rest
pandas
jinja2
psycopg2-binary
python-dotenv
gunicorn